Split ControlProblemElements.* in multiple files
